Add files via upload
Python code and Data set for stock market price prediction and forecasting using stacked LSTM. There are many columns like high,low,close,open for stock prediction and all are time series data.1st we have to take one column on which we want to forecast the next 30 days prediction.As this is a time series data we can't do train test split randomly  we have to do it manually. Forecasting using stacked LSTM is awesome.